Generate marshalling for datasource config builders
