+++ pkg/acceptance/bettertestspoc/config/datasourcemodel/database_model_gen.go
@@ -3,6 +3,8 @@
 package datasourcemodel
 
 import (
+	"encoding/json"
+
 	tfconfig "github.com/hashicorp/terraform-plugin-testing/config"
 
 	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/acceptance/bettertestspoc/config"
@@ -41,6 +43,26 @@ func DatabaseWithDefaultMeta(
 ) *DatabaseModel {
 	d := &DatabaseModel{DatasourceModelMeta: config.DatasourceDefaultMeta(datasources.Database)}
 	d.WithName(name)
+	return d
+}
+
+///////////////////////////////////////////////////////
+// set proper json marshalling and handle depends on //
+///////////////////////////////////////////////////////
+
+func (d *DatabaseModel) MarshalJSON() ([]byte, error) {
+	type Alias DatabaseModel
+	return json.Marshal(&struct {
+		*Alias
+		DependsOn []string `json:"depends_on,omitempty"`
+	}{
+		Alias:     (*Alias)(d),
+		DependsOn: d.DependsOn(),
+	})
+}
+
+func (d *DatabaseModel) WithDependsOn(values ...string) *DatabaseModel {
+	d.SetDependsOn(values...)
 	return d
 }
